+++ openid-data-classes/src/commonTest/kotlin/at/asitplus/openid/dcql/DCQLIsoMdocClaimsQueryTest.kt
@@ -1,6 +1,5 @@
 package at.asitplus.openid.dcql
 
-import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import at.asitplus.testballoon.invoke
 import at.asitplus.testballoon.minus
@@ -13,15 +12,18 @@ import io.ktor.util.*
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonPrimitive
+import kotlinx.serialization.json.add
+import kotlinx.serialization.json.buildJsonArray
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.decodeFromJsonElement
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlin.random.Random
 
+@Suppress("unused")
 val DCQLIsoMdocClaimsQueryTest by testSuite {
     "specification" - {
         "serial names" {
-            DCQLJsonClaimsQuery.SerialNames.PATH shouldBe "path"
+            DCQLClaimsQuery.SerialNames.PATH shouldBe "path"
         }
     }
     "instance serialization" {
@@ -34,23 +36,27 @@ val DCQLIsoMdocClaimsQueryTest by testSuite {
 
         val value = DCQLIsoMdocClaimsQuery(
             id = id,
-            namespace = namespace,
-            claimName = claimName,
+            path = DCQLClaimsPathPointer(
+                namespace,
+                claimName
+            ),
             values = values,
         )
 
         val correspondingJsonElement = buildJsonObject {
             put(DCQLClaimsQuery.SerialNames.ID, JsonPrimitive(id.string))
             put(DCQLClaimsQuery.SerialNames.VALUES, Json.encodeToJsonElement(values))
-            put(DCQLIsoMdocClaimsQuery.SerialNames.CLAIM_NAME, JsonPrimitive(claimName))
-            put(DCQLIsoMdocClaimsQuery.SerialNames.NAMESPACE, JsonPrimitive(namespace))
+            put(DCQLClaimsQuery.SerialNames.PATH, buildJsonArray {
+                add(namespace)
+                add(claimName)
+            })
         }
 
         val base: DCQLClaimsQuery = value
         val serialized = Json.encodeToJsonElement(base)
 
         serialized shouldBe correspondingJsonElement
-        Json.decodeFromJsonElement<DCQLClaimsQuery>(correspondingJsonElement) shouldBe base
+        Json.decodeFromJsonElement<DCQLIsoMdocClaimsQuery>(correspondingJsonElement) shouldBe base
     }
 
     "execution" {
@@ -68,120 +74,37 @@ val DCQLIsoMdocClaimsQueryTest by testSuite {
 
         DCQLIsoMdocClaimsQuery(
             id = DCQLClaimsQueryIdentifier("test"),
-            namespace = "testNamespace",
-            claimName = "testClaimName",
+            path = DCQLClaimsPathPointer(
+                "testNamespace",
+                "testClaimName",
+            ),
             values = listOf(
                 DCQLExpectedClaimValue.StringValue("test"),
                 DCQLExpectedClaimValue.IntegerValue(0),
                 DCQLExpectedClaimValue.BooleanValue(true),
             )
         ).executeIsoMdocClaimsQueryAgainstCredential(
-            credential = credential,
-            credentialStructureExtractor = {
-                DCQLCredentialClaimStructure.IsoMdocStructure(it)
-            },
-            credentialQuery = DCQLIsoMdocCredentialQuery(
-                id = DCQLCredentialQueryIdentifier(
-                    Random.nextBytes(32).encodeToString(Base64UrlStrict),
-                ),
-                format = CredentialFormatEnum.MSO_MDOC,
-                meta = DCQLIsoMdocCredentialMetadataAndValidityConstraints(
-                    doctypeValue = "anything"
-                )
-            )
+            credentialStructure = DCQLCredentialClaimStructure.IsoMdocStructure(credential)
         ).getOrThrow().shouldBeInstanceOf<DCQLClaimsQueryResult.IsoMdocResult>().let {
             it.claimValue shouldBe 0
         }
 
         DCQLIsoMdocClaimsQuery(
             id = DCQLClaimsQueryIdentifier("test"),
-            namespace = "testNamespace",
-            claimName = "testClaimName",
+            path = DCQLClaimsPathPointer(
+                "testNamespace",
+                "testClaimName",
+            ),
             values = listOf(
                 DCQLExpectedClaimValue.StringValue("test"),
                 DCQLExpectedClaimValue.IntegerValue(-1),
                 DCQLExpectedClaimValue.BooleanValue(true),
-            )
-        ).executeIsoMdocClaimsQueryAgainstCredential(
-            credential = credential,
-            credentialStructureExtractor = {
-                DCQLCredentialClaimStructure.IsoMdocStructure(it)
-            },
-            credentialQuery = DCQLIsoMdocCredentialQuery(
-                id = DCQLCredentialQueryIdentifier(
-                    Random.nextBytes(32).encodeToString(Base64UrlStrict),
-                ),
-                format = CredentialFormatEnum.MSO_MDOC,
-                meta = DCQLIsoMdocCredentialMetadataAndValidityConstraints(
-                    doctypeValue = "anything"
-                )
-            )
-        ).isSuccess shouldBe false
-
-        DCQLIsoMdocClaimsQuery(
-            id = DCQLClaimsQueryIdentifier("test"),
-            values = listOf(
-                DCQLExpectedClaimValue.StringValue("test"),
-                DCQLExpectedClaimValue.IntegerValue(0),
-                DCQLExpectedClaimValue.BooleanValue(true),
             ),
-            namespace = "testNamespace",
-            claimName = "testClaimName",
         ).executeIsoMdocClaimsQueryAgainstCredential(
-            credential = credential,
-            credentialStructureExtractor = {
-                DCQLCredentialClaimStructure.IsoMdocStructure(it)
-            },
-            credentialQuery = DCQLSdJwtCredentialQuery(
-                id = DCQLCredentialQueryIdentifier(
-                    Random.nextBytes(32).encodeToString(Base64UrlStrict),
-                ),
-                format = CredentialFormatEnum.DC_SD_JWT,
-                meta = DCQLSdJwtCredentialMetadataAndValidityConstraints(
-                    vctValues = listOf("anything")
-                )
-            )
+            credentialStructure = DCQLCredentialClaimStructure.IsoMdocStructure(credential)
         ).isSuccess shouldBe false
     }
 
-    "new and old parameters need to match" {
-        val namespace = uuid4().toString()
-        val claim = uuid4().toString()
-        val identifier = DCQLClaimsQueryIdentifier(uuid4().toString())
-        DCQLIsoMdocClaimsQuery(
-            id = identifier,
-            namespace = namespace,
-            claimName = claim,
-            path = DCQLClaimsPathPointer(namespace, claim)
-        )
-
-        shouldThrow<IllegalArgumentException> {
-            DCQLIsoMdocClaimsQuery(
-                id = identifier,
-                namespace = namespace,
-                claimName = claim,
-                path = DCQLClaimsPathPointer(namespace.reversed(), claim)
-            )
-        }
-
-        shouldThrow<IllegalArgumentException> {
-            DCQLIsoMdocClaimsQuery(
-                id = identifier,
-                namespace = namespace,
-                claimName = claim,
-                path = DCQLClaimsPathPointer(namespace, claim.reversed())
-            )
-        }
-
-        shouldThrow<IllegalArgumentException> {
-            DCQLIsoMdocClaimsQuery(
-                id = identifier,
-                namespace = namespace,
-                claimName = claim,
-                path = DCQLClaimsPathPointer(claim, namespace)
-            )
-        }
-    }
 
     "needs to have exactly two path items" {
         val namespace = uuid4().toString()