+++ openid-data-classes/src/commonTest/kotlin/at/asitplus/openid/dcql/DCQLQueryTest.kt
@@ -1,6 +1,5 @@
 package at.asitplus.openid.dcql
 
-import at.asitplus.csc.or
 import at.asitplus.data.NonEmptyList.Companion.nonEmptyListOf
 import at.asitplus.data.NonEmptyList.Companion.toNonEmptyList
 import at.asitplus.openid.CredentialFormatEnum
@@ -11,8 +10,6 @@ import at.asitplus.testballoon.withData
 import at.asitplus.testballoon.withDataSuites
 import com.benasher44.uuid.uuid4
 import de.infix.testBalloon.framework.core.testSuite
-import io.kotest.assertions.throwables.shouldNotThrowAny
-import io.kotest.assertions.throwables.shouldThrowAny
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.maps.shouldHaveSize
@@ -73,9 +70,9 @@ val DCQLQueryTest by testSuite {
                     "vct_values": [ "https://credentials.example.com/identity_credential" ]
                   },
                   "claims": [
-                      {"path": ["last_name"]},
-                      {"path": ["first_name"]},
-                      {"path": ["address", "street_address"]}
+                    {"path": ["last_name"] },
+                    {"path": ["first_name"] },
+                    {"path": ["address", "street_address"] }
                   ]
                 }
               ]
@@ -133,7 +130,9 @@ val DCQLQueryTest by testSuite {
                         ),
                     ),
                 ) {
-                    TestCredentialQueryAdapter(dcqlQuery).isSatisfiable(it) shouldBe false
+                    TestCredentialQueryAdapter(dcqlQuery).isSatisfiable(
+                        it,
+                    ) shouldBe false
                 }
             }
 
@@ -236,12 +235,10 @@ val DCQLQueryTest by testSuite {
                       },
                       "claims": [
                         {
-                          "namespace": "org.iso.7367.1",
-                          "claim_name": "vehicle_holder"
+                          "${DCQLClaimsQuery.SerialNames.PATH}": ["org.iso.7367.1", "vehicle_holder"]
                         },
                         {
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "first_name"
+                          "${DCQLClaimsQuery.SerialNames.PATH}": ["org.iso.18013.5.1", "first_name"]
                         }
                       ]
                     }
@@ -415,12 +412,10 @@ val DCQLQueryTest by testSuite {
                       },
                       "claims": [
                         {
-                          "namespace": "org.iso.7367.1",
-                          "claim_name": "vehicle_holder"
+                          "path": ["org.iso.7367.1", "vehicle_holder"]
                         },
                         {
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "first_name"
+                          "path": ["org.iso.18013.5.1", "first_name"]
                         }
                       ]
                     }
@@ -703,18 +698,15 @@ val DCQLQueryTest by testSuite {
                       "claims": [
                         {
                           "id": "given_name",
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "given_name"
+                          "path": ["org.iso.18013.5.1", "given_name"]
                         },
                         {
                           "id": "family_name",
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "family_name"
+                          "path": ["org.iso.18013.5.1", "family_name"]
                         },
                         {
                           "id": "portrait",
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "portrait"
+                          "path": ["org.iso.18013.5.1", "portrait"]
                         }
                       ]
                     },
@@ -727,13 +719,11 @@ val DCQLQueryTest by testSuite {
                       "claims": [
                         {
                           "id": "resident_address",
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "resident_address"
+                          "path": ["org.iso.18013.5.1", "resident_address"]
                         },
                         {
                           "id": "resident_country",
-                          "namespace": "org.iso.18013.5.1",
-                          "claim_name": "resident_country"
+                          "path": ["org.iso.18013.5.1", "resident_country"]
                         }
                       ]
                     },
@@ -746,18 +736,15 @@ val DCQLQueryTest by testSuite {
                       "claims": [
                         {
                           "id": "given_name",
-                          "namespace": "org.iso.23220.1",
-                          "claim_name": "given_name"
+                          "path": ["org.iso.23220.1", "given_name"]
                         },
                         {
                           "id": "family_name",
-                          "namespace": "org.iso.23220.1",
-                          "claim_name": "family_name"
+                          "path": ["org.iso.23220.1", "family_name"]
                         },
                         {
                           "id": "portrait",
-                          "namespace": "org.iso.23220.1",
-                          "claim_name": "portrait"
+                          "path": ["org.iso.23220.1", "portrait"]
                         }
                       ]
                     },
@@ -770,13 +757,11 @@ val DCQLQueryTest by testSuite {
                       "claims": [
                         {
                           "id": "resident_address",
-                          "namespace": "org.iso.23220.1",
-                          "claim_name": "resident_address"
+                          "path": ["org.iso.23220.1", "resident_address"]
                         },
                         {
                           "id": "resident_country",
-                          "namespace": "org.iso.23220.1",
-                          "claim_name": "resident_country"
+                          "path": ["org.iso.23220.1", "resident_country"]
                         }
                       ]
                     }
@@ -1397,8 +1382,7 @@ val DCQLQueryTest by testSuite {
                           "format": "${CredentialFormatEnum.MSO_MDOC.text}",
                           "claims": [
                               {
-                                "${DCQLIsoMdocClaimsQuery.SerialNames.NAMESPACE}": "namespace",
-                                "${DCQLIsoMdocClaimsQuery.SerialNames.CLAIM_NAME}": "claimName",
+                                "${DCQLClaimsQuery.SerialNames.PATH}": ["namespace", "claimName"],
                                 "${DCQLClaimsQuery.SerialNames.VALUES}": ${testVector.first}
                               }
                           ],