+++ openid-data-classes/src/commonTest/kotlin/at/asitplus/openid/dcql/DCQLSdJwtCredentialQueryTest.kt
@@ -42,7 +42,7 @@ val DCQLSdJwtCredentialQueryTest by testSuite {
                 })
             put(DCQLCredentialQuery.SerialNames.CLAIMS, buildJsonArray {
                 add(buildJsonObject {
-                    put(DCQLJsonClaimsQuery.SerialNames.PATH, buildJsonArray {
+                    put(DCQLClaimsQuery.SerialNames.PATH, buildJsonArray {
                         add(JsonNull)
                     })
                 })